+++ multiboot2-header/Changelog.md
@@ -3,7 +3,8 @@
 ## Unreleased
 
 - added `EndHeaderTag::default()`
-- MSRV is 1.69
+- MSRV is 1.70
+- Can add multiple `TagType::Smbios` tags in the builder.
 
 ## 0.3.2 (2023-11-30)